+++ packages/core/src/languages/html/css-properties.ts
@@ -1,7 +1,10 @@
 import type { MangleExpression } from "../../types";
 import type { CssDeclarationPropertyOptions } from "../options";
 
-import { NestedGroupMangleExpression } from "../utils/mangle-expressions";
+import {
+  NestedGroupMangleExpression,
+  SingleGroupMangleExpression,
+} from "../utils/mangle-expressions";
 import { QUOTED_ATTRIBUTE_PATTERN, QUOTES_ARRAY } from "./common";
 
 const GROUP_MAIN = "main";
@@ -59,6 +62,36 @@ function newStyleDeclarationPropertyExpressions(
 }
 
 /**
+ * Get {@link MangleExpression}s to match the property of CSS declarations in
+ * unquoted HTML attributes, e.g. `font` in `<div style=font:serif><div>"`.
+ *
+ * @param propertyPrefix The prefix required on properties.
+ * @param propertySuffix The suffix required on properties.
+ * @returns The {@link MangleExpression}s to match unquoted attribute values.
+ */
+function newUnquotedStyleDeclarationPropertyExpressions(
+  propertyPrefix: string,
+  propertySuffix: string,
+): MangleExpression[] {
+  return QUOTES_ARRAY.map((quote) => new SingleGroupMangleExpression(
+    `
+      (?<=
+        \\<\\s*[a-zA-Z0-9]+\\s+
+        (?:
+          [^>\\s=]+
+          (?:\\s*=\\s*${quote}[^${quote}]*${quote})?
+          \\s+
+        )*
+        style\\s*=\\s*${propertyPrefix}
+      )
+      (?<${GROUP_MAIN}>%s)
+      (?=${propertySuffix}:)
+    `,
+    GROUP_MAIN,
+  ));
+}
+
+/**
  * Get the set of {@link MangleExpression}s to match the properties of CSS
  * declarations in HTML. This will match:
  * - Properties in style attributes (e.g. `foo` in `<img style="foo: bar">`).
@@ -66,7 +99,7 @@ function newStyleDeclarationPropertyExpressions(
  * @param options The {@link CssDeclarationPropertyOptions}.
  * @returns A set of {@link MangleExpression}s.
  * @since v0.1.14
- * @version v0.1.18
+ * @version v0.1.19
  */
 export default function cssDeclarationPropertyExpressionFactory(
   options: CssDeclarationPropertyOptions,
@@ -76,5 +109,9 @@ export default function cssDeclarationPropertyExpressionFactory(
 
   return [
     ...newStyleDeclarationPropertyExpressions(propertyPrefix, propertySuffix),
+    ...newUnquotedStyleDeclarationPropertyExpressions(
+      propertyPrefix,
+      propertySuffix,
+    ),
   ];
 }